correct misreading outputs in MajorityElement.java change all ints to long in Inversions.java as some inputs are larger then 10^9

+++ algorithmic-toolbox/src/main/java/week4/assignement/Inversions.java
@@ -4,11 +4,11 @@ import java.util.Scanner;
 
 public class Inversions {
 
-    protected static long getNumberOfInversions(int[] arr) {
+    protected static long getNumberOfInversions(long[] arr) {
         return sortAndCountInversions(arr, 0, arr.length - 1);
     }
 
-    private static long sortAndCountInversions(int[] arr, int left, int right) {
+    private static long sortAndCountInversions(long[] arr, int left, int right) {
         long inversions = 0;
         if (right > left) {
             int mid = (right + left) / 2;
@@ -20,16 +20,16 @@ public class Inversions {
         return inversions;
     }
 
-    private static long merge(int[] arr, int left, int mid, int right) {
+    private static long merge(long[] arr, int left, int mid, int right) {
         // pointer in first array [from left to mid - 1]
         int i = left;
         // pointer in second array [from mid to right]
         int j = mid;
         // store sorted values in temp array
-        int[] temp = new int[right - left + 1];
+        long[] temp = new long[right - left + 1];
         // index in temp array
         int k = 0;
-        int inversions = 0;
+        long inversions = 0;
         // loop and sort while counting inversions
         while ((i <= mid - 1) && (j <= right)) {
             if (arr[i] <= arr[j]) {
@@ -57,7 +57,7 @@ public class Inversions {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         int n = scanner.nextInt();
-        int[] a = new int[n];
+        long[] a = new long[n];
         for (int i = 0; i < n; i++) {
             a[i] = scanner.nextInt();
         }
